modulegen: create internal/make
