+++ modulegen/_template/Makefile.tmpl
@@ -1,5 +1,5 @@
-{{ $lower := ToLower }}include ../../commons-test.mk
+include ../../commons-test.mk
 
 .PHONY: test
 test:
-	$(MAKE) test-{{ $lower }}
+	$(MAKE) test-{{ . }}
